la création des entités et des DTOs nécessaires pour gérer et transférer les données dans l'applicatio

+++ src/main/java/com/example/rh/Models/Dtos/Request/ApplicationRequestDTO.java
@@ -11,12 +11,14 @@ public class ApplicationRequestDTO {
     private String candidateCin;
 
     @NotNull(message = "Offer is mandatory")
-    private Integer offerId;
+    private Long offerId;
 
     private String etatCandidature;
     public Application toModel() {
         Candidate candidate = Candidate.builder().cin(candidateCin).build();
+
         Offer offer = Offer.builder().id(offerId).build();
+
         EtatCandidature etatCandidatureEnum = EtatCandidature.valueOf(etatCandidature);
 
         Application application = Application.builder()
